Move new configuration packages and add test for PropertyMap
