+++ src/main/java/fr/xephi/authme/settings/domain/Comment.java
@@ -1,4 +1,4 @@
-package fr.xephi.authme.settings.custom.domain;
+package fr.xephi.authme.settings.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
