+++ src/main/java/fr/xephi/authme/settings/domain/PropertyType.java
@@ -1,4 +1,4 @@
-package fr.xephi.authme.settings.custom.domain;
+package fr.xephi.authme.settings.domain;
 
 import org.bukkit.configuration.file.YamlConfiguration;
